# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "GIN CRUD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "assusa456u.local:8802",
    "basePath": "/",
    "paths": {
...
        "/v1/t_token": {
            "get": {
                "description": "Get Page TToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenPage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "0",
                        "description": "page",
                        "name": "_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "5",
                        "description": "size",
                        "name": "_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "_sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter",
                        "name": "_filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "global filter",
                        "name": "_q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create TToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenCreate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Add TToken",
                        "name": "tToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/t_token/delete/{id}": {
            "put": {
                "description": "Soft Delete TToken by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenSoftDelete",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "TToken id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/t_token/header": {
            "get": {
                "description": "Get TToken header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenHeader",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/t_token/{id}": {
            "get": {
                "description": "Get TToken by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenIndex",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "TToken id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update TToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Update TToken",
                        "name": "tToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TToken"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "TToken id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete TToken by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tToken"
                ],
                "summary": "TTokenDelete",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "TToken id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.MBiodata": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "fullname": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "imagePath": {
                    "type": "string",
                    "maxLength": 255
                },
                "isDelete": {
                    "type": "boolean"
                },
                "mobilePhone": {
                    "type": "string",
                    "maxLength": 15
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                }
            }
        },
...
        "model.MUser": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "biodataId": {
                    "type": "integer"
                },
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "id": {
                    "type": "integer"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "isLocked": {
                    "type": "boolean"
                },
                "lastLogin": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "loginAttempt": {
                    "type": "integer"
                },
                "mbiodata": {
                    "$ref": "#/definitions/model.MBiodata"
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "mrole": {
                    "$ref": "#/definitions/model.MRole"
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
...
        "model.TToken": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "expiredOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "id": {
                    "type": "integer"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "isExpired": {
                    "type": "boolean"
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "muser": {
                    "$ref": "#/definitions/model.MUser"
                },
                "token": {
                    "type": "string",
                    "maxLength": 50
                },
                "usedFor": {
                    "type": "string",
                    "maxLength": 20
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "path": {
                    "type": "string",
                    "example": "/v1/m_biodata"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}